dom/tests/mochitest/geolocation/test_manyWatchSerial.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=482260
-->
<head>
  <title>Test for watchPosition </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=482260">Mozilla Bug 482260</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

var watchID = 0;
var completeCount = 10;


function successCallback(position) {

  navigator.geolocation.clearWatch(watchID);

  completeCount--;

  if (completeCount==0) {
    ok(1, "all watchPosition successCallbacks called");
    SimpleTest.finish();
    return;
  }

  watchID = navigator.geolocation.watchPosition(successCallback);
}

function test1() {
  watchID = navigator.geolocation.watchPosition(successCallback);
}
</script>
</pre>
</body>
</html>
